MyProject/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.13.0
drf-extra-fields==3.4.1
flake8==6.1.0
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.5.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
python-decouple==3.5
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.3.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
